# internal/features/auth/dto.go
package auth

import "financial_control/internal/features/user"

type LoginRequest struct {
	Email       string `json:"email"`
	Password    string `json:"password"`
	GoogleToken string `json:"googleToken,omitempty"`
}

type LoginResponse struct {
	AccessToken  string            `json:"access_token"`
	RefreshToken string            `json:"refresh_token"`
	ExpiresIn    int64             `json:"expires_in"`
	User         user.UserResponse `json:"user"`
}

type RefreshRequest struct {
	RefreshToken string `json:"refresh_token"`
}

type PasswordRequest struct {
	CurrentPassword string `json:"currentPassword"`
	NewPassword     string `json:"newPassword"`
	ConfirmPassword string `json:"confirmPassword"`
}